(fix): CI/CD setup #5
